requirements.txt
# roughly equivalent to `python setup.py develop`
-e .

astroid==2.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
getch==1.0
gitdb==4.0.9
github3.py==3.0.0
GitPython==3.1.24
idna==2.10
isort==5.7.0
jwcrypto==0.8
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.22.1
pycparser==2.20
pylint==2.7.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
smmap==5.0.0
toml==0.10.2
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
wrapt==1.12.1
# alive_progress==2.4.1
# MoCo-GHE-Admin's fork of aerickson's fork of alive_progress 2.4.1 that writes to stderr (vs stdout)
# - switch back to the released version when https://github.com/rsalmei/alive-progress/issues/177 is resolved
git+https://github.com/MoCo-GHE-Admin/alive-progress@02e8eb93#egg=alive_progress
